refactor(daemon): extract pointer turn runner, drop setPointerMessageProcessor DI
The pointer-message processor was injected into calls/call-pointer-messages.ts
via a module-level setter (setPointerMessageProcessor) wired from lifecycle.ts
with a ~100-line inline closure. The closure captured nothing from the lifecycle
scope — it only used module-level imports — so it is freely extractable.

- Extract the closure into daemon/pointer-turn-runner.ts as
  runPointerMessageTurn(). It owns the conversation-turn orchestration
  (trust elevation, tool-disabled agent loop, fact-check, rollback).
- call-pointer-messages.ts imports runPointerMessageTurn directly and calls it
  on the trusted-audience path; the existing try/catch still falls back to
  deterministic copy on failure. Removes the PointerMessageProcessor type,
  the module-level slot, and the set/reset setters.
- lifecycle.ts drops the injection block and its now-unused imports; keeps the
  post-HTTP broadcastDaemonStatus() call.

Cycle note: this is a module extraction (not an inline/dynamic import).
call-pointer-messages was already inside the conversation SCC via its existing
conversation-crud import, so no module that was previously cycle-free gains a
new SCC; the runner joins the same component, and every cross-module reference
is used at call time (not module-eval), so init order is unaffected.

Test wires the mocked runner via a swappable impl instead of the removed setter.
