+++ assistant/src/__tests__/call-pointer-messages.test.ts
@@ -7,11 +7,40 @@ mock.module("../util/logger.js", () => ({
     }),
 }));
 
+// Stand in for the daemon conversation-turn path so these tests can drive its
+// behavior (success / throw) without pulling in the real conversation pipeline.
+// `runPointerTurnImpl` is swapped per-test via setProcessor/resetProcessor.
+let runPointerTurnImpl: (
+  conversationId: string,
+  instruction: string,
+  requiredFacts?: string[],
+) => Promise<void> = async () => {};
+
+mock.module("../daemon/pointer-turn-runner.js", () => ({
+  runPointerMessageTurn: (
+    conversationId: string,
+    instruction: string,
+    requiredFacts?: string[],
+  ) => runPointerTurnImpl(conversationId, instruction, requiredFacts),
+}));
+
+function setProcessor(
+  fn: (
+    conversationId: string,
+    instruction: string,
+    requiredFacts?: string[],
+  ) => Promise<void>,
+): void {
+  runPointerTurnImpl = fn;
+}
+
+function resetProcessor(): void {
+  runPointerTurnImpl = async () => {};
+}
+
 import {
   addPointerMessage,
   formatDuration,
-  resetPointerMessageProcessor,
-  setPointerMessageProcessor,
 } from "../calls/call-pointer-messages.js";
 import { addMessage, getMessages } from "../persistence/conversation-crud.js";
 import { getDb } from "../persistence/db-connection.js";
@@ -97,7 +126,7 @@ describe("addPointerMessage", () => {
   });
 
   afterEach(() => {
-    resetPointerMessageProcessor();
+    resetProcessor();
   });
 
   test("adds a started pointer message", () => {
@@ -204,7 +233,7 @@ describe("addPointerMessage", () => {
     ensureConversation(convId);
 
     const processorCalled = { value: false };
-    setPointerMessageProcessor(async () => {
+    setProcessor(async () => {
       processorCalled.value = true;
     });
 
@@ -220,7 +249,7 @@ describe("addPointerMessage", () => {
     ensureConversation(convId, { originChannel: "vellum" });
 
     const processorCalled = { value: false };
-    setPointerMessageProcessor(async () => {
+    setProcessor(async () => {
       processorCalled.value = true;
     });
 
@@ -242,7 +271,7 @@ describe("addPointerMessage", () => {
 
     let capturedInstruction = "";
     let capturedFacts: string[] = [];
-    setPointerMessageProcessor(async (_convId, instruction, requiredFacts) => {
+    setProcessor(async (_convId, instruction, requiredFacts) => {
       capturedInstruction = instruction;
       capturedFacts = requiredFacts ?? [];
     });
@@ -265,7 +294,7 @@ describe("addPointerMessage", () => {
     const convId = "conv-ptr-processor-fail";
     ensureConversation(convId, { originChannel: "vellum" });
 
-    setPointerMessageProcessor(async () => {
+    setProcessor(async () => {
       throw new Error("Daemon unavailable");
     });
 
@@ -282,7 +311,7 @@ describe("addPointerMessage", () => {
     ensureConversation(convId, { originChannel: "vellum" });
 
     let processorCalled = false;
-    setPointerMessageProcessor(async () => {
+    setProcessor(async () => {
       processorCalled = true;
     });
 
@@ -297,7 +326,7 @@ describe("addPointerMessage", () => {
     ensureConversation(convId);
 
     const processorCalled = { value: false };
-    setPointerMessageProcessor(async () => {
+    setProcessor(async () => {
       processorCalled.value = true;
     });
 
@@ -318,7 +347,7 @@ describe("addPointerMessage", () => {
     });
 
     let processorCalled = false;
-    setPointerMessageProcessor(async () => {
+    setProcessor(async () => {
       processorCalled = true;
     });
 
@@ -335,7 +364,7 @@ describe("addPointerMessage", () => {
     });
 
     let processorCalled = false;
-    setPointerMessageProcessor(async () => {
+    setProcessor(async () => {
       processorCalled = true;
     });
 
@@ -362,7 +391,7 @@ describe("addPointerMessage", () => {
 
     // And that pointer-audience trust treats it identically to trusted_contact.
     let processorCalled = false;
-    setPointerMessageProcessor(async () => {
+    setProcessor(async () => {
       processorCalled = true;
     });
 
@@ -378,7 +407,7 @@ describe("addPointerMessage", () => {
     });
 
     const processorCalled = { value: false };
-    setPointerMessageProcessor(async () => {
+    setProcessor(async () => {
       processorCalled.value = true;
     });
