+++ assistant/src/daemon/lifecycle.ts
@@ -1,6 +1,5 @@
 import { config as dotenvConfig } from "dotenv";
 
-import { setPointerMessageProcessor } from "../calls/call-pointer-messages.js";
 import { reconcileCallsOnStartup } from "../calls/call-recovery.js";
 import { TwilioConversationRelayProvider } from "../calls/twilio-provider.js";
 import { initFeatureFlagOverrides } from "../config/assistant-feature-flags.js";
@@ -18,11 +17,7 @@ import { startCliIpcServer } from "../ipc/assistant-server.js";
 import { startGatewayFlagListener } from "../ipc/gateway-flag-listener.js";
 import { backfillManualTokenConnections } from "../oauth/manual-token-connection.js";
 import { seedOAuthProviders } from "../oauth/seed-providers.js";
-import {
-  clearStaleProcessingFlags,
-  deleteMessageById,
-  getMessages,
-} from "../persistence/conversation-crud.js";
+import { clearStaleProcessingFlags } from "../persistence/conversation-crud.js";
 import { getDb } from "../persistence/db-connection.js";
 import { initializeDb } from "../persistence/db-init.js";
 import { startEmbeddingRuntimeManager } from "../persistence/embeddings/embedding-runtime-manager.js";
@@ -66,7 +61,6 @@ import { runWorkspaceMigrations } from "../workspace/migrations/runner.js";
 import { startAppSourceWatcher } from "./app-source-watcher.js";
 import { startConfigWatcher } from "./config-watcher.js";
 import { startConversationEvictor } from "./conversation-evictor.js";
-import { getOrCreateConversation } from "./conversation-store.js";
 import { writePid } from "./daemon-control.js";
 import { setDbReady, setStartupComplete } from "./daemon-readiness.js";
 import {
@@ -79,7 +73,6 @@ import { initializePlugins } from "./external-plugins-bootstrap.js";
 import { backfillSlackInjectionTemplates } from "./handlers/config-slack-channel.js";
 import { installAssistantSymlink } from "./install-symlink.js";
 import { startOrphanReaper } from "./orphan-reaper.js";
-import { elevatePointerConversationToGuardian } from "./pointer-conversation-trust.js";
 import { runProfilerSweep } from "./profiler-run-store.js";
 import {
   initializeProvidersAndTools,
@@ -667,156 +660,9 @@ export async function runDaemon(): Promise<void> {
     log.warn({ err }, "Background Qdrant init failed"),
   );
 
-  try {
-    setPointerMessageProcessor(
-      async (conversationId, instruction, requiredFacts) => {
-        const conversation = await getOrCreateConversation(conversationId);
-
-        // Pointer turns are guardian-gated owner self-maintenance: elevate to
-        // the internal guardian context and rehydrate history so a cold
-        // (evicted) load doesn't filter guardian history to empty and ship a
-        // cache-missing turn. `restoreTrustContext` undoes the elevation after
-        // the turn. See pointer-conversation-trust.ts for the full rationale.
-        const restoreTrustContext =
-          await elevatePointerConversationToGuardian(conversation);
-
-        // Constrain pointer generation to a tool-disabled path so call-
-        // status events cannot trigger unintended side-effect tools.
-        // Incrementing toolsDisabledDepth causes the resolveTools callback
-        // to return an empty tool list, preventing the LLM from seeing or
-        // invoking any tools during the pointer agent loop.
-        //
-        // A depth counter (rather than a boolean) ensures that overlapping
-        // pointer requests on the same conversation don't clear each other's
-        // constraint — each caller increments on entry and decrements in
-        // its own finally block.
-        conversation.toolsDisabledDepth++;
-        try {
-          const { id: messageId } = await conversation.persistUserMessage({
-            content: instruction,
-            metadata: { pointerInstruction: true },
-            displayContent: "[Call status event]",
-          });
-
-          // Helper: roll back persisted messages on failure, then reload
-          // in-memory history from the (now cleaned) DB. Reloading avoids
-          // stale-index issues when context compaction reassigns the
-          // messages array during runAgentLoop.
-          const rollback = async (extraMessageIds?: string[]) => {
-            try {
-              deleteMessageById(messageId);
-            } catch {
-              /* best effort */
-            }
-            for (const id of extraMessageIds ?? []) {
-              try {
-                deleteMessageById(id);
-              } catch {
-                /* best effort */
-              }
-            }
-            try {
-              await conversation.loadFromDb();
-            } catch {
-              /* best effort */
-            }
-          };
-
-          // Snapshot message IDs before the agent loop so we can diff
-          // afterwards to find exactly which messages this run created,
-          // avoiding positional heuristics that break under concurrency.
-          //
-          // Caveat: the diff captures *all* new messages in the
-          // conversation during the loop window, not just those from
-          // this specific agent loop.  If a concurrent pointer event
-          // falls back to a deterministic addMessage() while our loop
-          // is in flight, that message lands in our diff.  The race
-          // requires two pointer events for the same conversation
-          // within the agent loop window *and* this run must fail or
-          // fail fact-check — narrow enough to accept.  A future
-          // improvement could tag messages with a per-run correlation
-          // ID so rollback only targets its own output.
-          const preRunMessageIds = new Set(
-            getMessages(conversationId).map((m) => m.id),
-          );
-
-          let agentLoopError: string | undefined;
-          let generatedText = "";
-          await conversation.runAgentLoop(instruction, messageId, {
-            onEvent: (msg) => {
-              if (
-                "type" in msg &&
-                msg.type === "assistant_text_delta" &&
-                "text" in msg
-              ) {
-                generatedText += (msg as { text: string }).text;
-              }
-              if (
-                "type" in msg &&
-                (msg.type === "error" || msg.type === "conversation_error")
-              ) {
-                agentLoopError =
-                  "message" in msg
-                    ? (msg as { message: string }).message
-                    : "userMessage" in msg
-                      ? (msg as { userMessage: string }).userMessage
-                      : "Agent loop failed";
-              }
-            },
-          });
-
-          // Identify messages created during this run by diffing against
-          // the pre-run snapshot. This captures all messages added to the
-          // conversation during the loop window, which may include messages
-          // from concurrent pointer events (see over-capture caveat above).
-          const postRunMessages = getMessages(conversationId);
-          const createdMessageIds = postRunMessages
-            .filter((m) => !preRunMessageIds.has(m.id) && m.id !== messageId)
-            .map((m) => m.id);
-
-          if (agentLoopError) {
-            await rollback(createdMessageIds);
-            throw new Error(agentLoopError);
-          }
-
-          // Post-generation fact check: verify the assistant's response
-          // includes all required factual details (phone number, duration,
-          // outcome keyword, etc.). If the model omitted or rewrote them,
-          // remove both the instruction and generated messages and throw so
-          // the deterministic fallback fires.
-          //
-          // Validation uses text accumulated from assistant_text_delta
-          // events during the agent loop rather than a DB lookup, avoiding
-          // any positional ambiguity when concurrent pointer events
-          // interleave messages in the conversation.
-          if (requiredFacts && requiredFacts.length > 0) {
-            const missingFacts = requiredFacts.filter(
-              (fact) => !generatedText.includes(fact),
-            );
-            if (missingFacts.length > 0) {
-              log.warn(
-                { conversationId, missingFacts },
-                "Generated pointer text failed fact validation — falling back to deterministic",
-              );
-              await rollback(createdMessageIds);
-              throw new Error("Generated pointer text failed fact validation");
-            }
-          }
-        } finally {
-          // Restore tool availability so subsequent turns aren't affected.
-          conversation.toolsDisabledDepth--;
-          // Undo the temporary guardian elevation installed above.
-          restoreTrustContext();
-        }
-      },
-    );
-    broadcastDaemonStatus();
-  } catch (err) {
-    log.warn(
-      { err },
-      "Failed to wire runtime HTTP server deps, continuing without them",
-    );
-  }
+  // The runtime HTTP server is up; broadcast the fresh daemon status so
+  // connected clients pick up the transition.
+  broadcastDaemonStatus();
 
   // Register built-in TTS providers so the provider abstraction can resolve
   // them by ID. Must happen before call controllers or routes are created.
